+++ M5-frontend/src/components/filterPage/Bed.jsx
