+++ M5-frontend/src/components/filterPage/Location.jsx
@@ -3,7 +3,6 @@ import TextField from '@mui/material/TextField';
 import Stack from '@mui/material/Stack';
 import Autocomplete from '@mui/material/Autocomplete';
 import axios from 'axios';
-import { list } from 'postcss';
 
 export default function Location({ onLocationChange }) {
   const [suburb, setSuburb] = useState([]);
