+++ M5-frontend/src/components/filterPage/Price.jsx
@@ -1,6 +1,6 @@
 import { useState } from 'react';
 import InputRange from 'react-input-range';
-import '../styles/slider-input-range.css';
+import '../../components/filterPage/slider-input-range.css';
 
 export default function Price({ onPriceChange }) {
   const [priceRange, setPriceRange] = useState({ min: 100, max: 1000 });
